requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1607439979954/work
aiohttp @ file:///C:/ci/aiohttp_1607109697839/work
altair==4.1.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828585465/work
asgiref==3.3.4
astor==0.8.1
astunparse==1.6.3
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
base58==2.1.0
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1611600262290/work
certifi==2022.6.15
cffi @ file:///C:/ci/cffi_1606255208697/work
chardet @ file:///C:/ci/chardet_1605303225733/work
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
commonmark==0.9.1
cryptography==2.9.2
cycler==0.10.0
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
defusedxml==0.6.0
Django==3.2
entrypoints==0.3
gast @ file:///tmp/build/80754af9/gast_1597433534803/work
gitdb==4.0.5
GitPython==3.1.12
google-auth @ file:///tmp/build/80754af9/google-auth_1607969906642/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1603929124518/work
google-pasta==0.2.0
grpcio @ file:///C:/ci/grpcio_1597406462198/work
h5py==2.10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imageio==2.13.3
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel @ file:///C:/ci/ipykernel_1596190155316/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1612212935303/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
joblib @ file:///tmp/build/80754af9/joblib_1607970656719/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213356021/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets==1.0.0
Keras @ file:///tmp/build/80754af9/keras_split_1593112142734/work
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.0
kiwisolver @ file:///C:/ci/kiwisolver_1604014703538/work
lxml==4.6.3
Markdown @ file:///C:/ci/markdown_1605111189761/work
MarkupSafe==1.1.1
matplotlib @ file:///C:/ci/matplotlib-base_1603355780617/work
missingno==0.5.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///C:/ci/multidict_1600456481656/work
multitasking==0.0.9
nb-conda==2.2.1
nb-conda-kernels @ file:///C:/ci/nb_conda_kernels_1606757941167/work
nbclient @ file:///tmp/build/80754af9/nbclient_1602783176460/work
nbconvert @ file:///C:/ci/nbconvert_1601914925608/work
nbformat @ file:///tmp/build/80754af9/nbformat_1610738111109/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1606153767164/work
networkx==2.6.3
nltk @ file:///home/conda/feedstock_root/build_artifacts/nltk_1618932666659/work
notebook @ file:///C:/ci/notebook_1611348396724/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603466732592/work
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python-headless==4.5.4.60
opt-einsum==3.1.0
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas @ file:///C:/ci/pandas_1611171485393/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102497129/work
parso @ file:///tmp/build/80754af9/parso_1607623074025/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
plotly==4.14.3
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1606344362066/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.13.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydeck==0.6.0
pydot==1.4.2
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
PyJWT @ file:///C:/ci/pyjwt_1610893382614/work
Pympler==1.0.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyreadline==2.1
pyrsistent @ file:///C:/ci/pyrsistent_1600141795814/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets==1.2.0
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
regex @ file:///D:/bld/regex_1617644574793/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-oauthlib==1.3.0
retrying==1.3.3
rich==12.4.4
rsa @ file:///tmp/build/80754af9/rsa_1610483308194/work
Sastrawi==1.0.1
scikit-image==0.19.1
scikit-learn @ file:///C:/ci/scikit-learn_1598377018496/work
scipy @ file:///C:/ci/scipy_1604596260408/work
seaborn @ file:///tmp/build/80754af9/seaborn_1608578541026/work
semver==2.13.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
sip==4.19.13
six @ file:///C:/ci/six_1605187374963/work
smmap==3.0.5
sqlparse==0.4.1
streamlit==1.10.0
tensorboard @ file:///home/builder/ktietz/conda/conda-bld/tensorboard_1604313476433/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator @ file:///tmp/build/80754af9/tensorflow-estimator_1599136169057/work/whl_temp/tensorflow_estimator-2.3.0-py2.py3-none-any.whl
termcolor==1.1.0
terminado==0.9.2
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
textblob==0.15.3
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2021.11.2
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
toolz==0.11.1
torch==1.4.0
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1623518955413/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
tweepy==3.10.0
typing_extensions==4.3.0
tzlocal==2.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1611694770489/work
validators==0.18.2
watchdog==1.0.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
widgetsnbextension==3.5.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wrapt==1.12.1
xgboost==1.4.2
yarl @ file:///C:/ci/yarl_1598045274898/work
yfinance==0.1.63
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
